update tf_2dunet workspace to use tensorflow2 aVnd remove legacy tensorflowV1 workspace
