+++ openfl-workspace/tf_2dunet/plan/data.yaml
@@ -1,8 +1,8 @@
-# Copyright (C) 2020-2021 Intel Corporation
+# Copyright (C) 2020-2024 Intel Corporation
 # Licensed subject to the terms of the separately executed evaluation license agreement between Intel Corporation and you.
 
 # all keys under 'collaborators' corresponds to a specific colaborator name the corresponding dictionary has data_name, data_path pairs.
 # Note that in the mnist case we do not store the data locally, and the data_path is used to pass an integer that helps the data object
 # construct the shard of the mnist dataset to be use for this collaborator.
-one,/raid/datasets/MICCAI_BraTS_2019_Data_Training/HGG/0
-two,/raid/datasets/MICCAI_BraTS_2019_Data_Training/HGG/1
+collaborator1, ~/MICCAI_BraTS_2019_Data_Training/HGG/0
+collaborator2, ~/MICCAI_BraTS_2019_Data_Training/HGG/1
