+++ openfl-workspace/tf_2dunet/plan/plan.yaml
@@ -1,4 +1,4 @@
-# Copyright (C) 2020-2021 Intel Corporation
+# Copyright (C) 2020-2024 Intel Corporation
 # Licensed subject to the terms of the separately executed evaluation license agreement between Intel Corporation and you.
 
 aggregator :
@@ -20,7 +20,7 @@ collaborator :
 
 data_loader :
   defaults : plan/defaults/data_loader.yaml
-  template : src.tfbrats_inmemory.TensorFlowBratsInMemory
+  template : src.dataloader.TensorFlowBratsInMemory
   settings :
     batch_size: 64
     percent_train: 0.8
@@ -29,7 +29,7 @@ data_loader :
 
 task_runner :
   defaults : plan/defaults/task_runner.yaml
-  template : src.tf_2dunet.TensorFlow2DUNet
+  template : src.taskrunner.TensorFlow2DUNet
 
 network :
   defaults : plan/defaults/network.yaml
@@ -37,8 +37,32 @@ network :
 assigner :
   defaults : plan/defaults/assigner.yaml
 
-tasks :
-  defaults : plan/defaults/tasks_tensorflow.yaml
+tasks:
+  defaults : plan/defaults/task_tensorflow.yaml
+  aggregated_model_validation:
+    function : validate_task
+    kwargs   :
+      batch_size : 32
+      apply      : global
+      metrics    :
+        - dice_coef
+        - soft_dice_coef
+  locally_tuned_model_validation:
+    function : validate_task
+    kwargs   :
+      batch_size : 32
+      apply      : local
+      metrics    :
+        - dice_coef
+        - soft_dice_coef
+  train:
+    function : train_task
+    kwargs   :
+      batch_size : 32
+      metrics    :
+      - loss
+      epochs     : 1
+
 
 compression_pipeline :
   defaults : plan/defaults/compression_pipeline.yaml
