+++ openfl-workspace/tf_2dunet/requirements.txt
@@ -1,3 +1,3 @@
 nibabel
-tensorflow==2.13
+tensorflow==2.15.1
 setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
